# set useApplyAsync for $http