According to following document, this option is useful if the application call many HTTP request.

> Configure $http service to combine processing of multiple http responses received at around the same time via $rootScope.$applyAsync. This can result in significant performance improvement for bigger applications that make many HTTP requests concurrently (common during application bootstrap).

https://docs.angularjs.org/api/ng/provider/$httpProvider

I test this and I confirm that $scope.digest calling count is significantly reduced.