@bergquist 2016-06-03
Nice! 

Digests goes from ~80 to ~50 on my comp on http://play.grafana.org/dashboard/db/big-dashboard