@mtanda 2016-06-04
My drilldown dashboard contains approximately 200 graph panels, and loading time is little slow.
I try to tuning loading process now!