@torkelo 2016-06-06
Nice! Been looking for more ways to limit digest during load! 

When you say you have 200 graph panels, do you have some in collapsed rows? collapsed rows does not cause any loads. 

I have started working on lazy loading panels as they are scrolled into view to better support long dashboards. 

https://github.com/grafana/grafana/tree/lazy_panels
https://github.com/grafana/grafana/issues/5216