@mtanda 2016-06-07
> When you say you have 200 graph panels, do you have some in collapsed rows? collapsed rows does not cause any loads.

Yes, I collapse rows, but I want to expand the rows if loading speed is enough fast.

This is the peformance improvement result in my env.
(use this to get result, https://github.com/grafana/grafana/pull/5286)

before:
```
Dashboard::Performance Total Digests: 365
Dashboard::Performance Total Watchers: 4447
Dashboard::Performance Total ScopeCount: 1044
Dashboard::Performance - All panels initialized in 745 ms
Dashboard::Performance Root Digest 3.0004999999999806
Dashboard::Performance - All panels rendered in 8347 ms
```

after:
```
Dashboard::Performance Total Digests: 19
Dashboard::Performance Total Watchers: 4447
Dashboard::Performance Total ScopeCount: 1044
Dashboard::Performance - All panels initialized in 949 ms
Dashboard::Performance - All panels rendered in 5120 ms
Dashboard::Performance Root Digest 43.2358333333333
```

8347ms => 5120ms, so the improvement is big!

> I have started working on lazy loading panels as they are scrolled into view to better support long dashboards.

Nice! It must be very useful :+1::+1: