@mtanda 2016-06-07
If we can remove a number of performance warning (like "Forced reflow is a likely performance bottleneck."), the loading speed might become more faster.

![force_reflow](https://cloud.githubusercontent.com/assets/224552/15845319/e57aeab2-2cac-11e6-8e4d-10f81b5b851a.png)